requirements.txt
# ─── Django Core ───────────────────────────────────────────────────────────────
Django==4.2.13
djangorestframework==3.15.1
django-environ==0.11.2
django-redis==5.4.0
django-ratelimit==4.1.0
django-storages==1.14.3
whitenoise==6.7.0

# ─── Auth & Security ────────────────────────────────────────────────────────────
django-allauth==0.63.3
django-cors-headers==4.4.0

# ─── Database ──────────────────────────────────────────────────────────────────
psycopg2-binary==2.9.9
django-extensions==3.2.3

# ─── Async / Task Queue ─────────────────────────────────────────────────────────
celery==5.4.0
redis==5.0.7
kombu==5.3.7
flower==2.0.1

# ─── AI / LangChain ─────────────────────────────────────────────────────────────
langchain==0.2.12
langchain-openai==0.1.20
langchain-pinecone==0.1.3
langchain-community==0.2.11
openai==1.37.1
tiktoken==0.7.0
pinecone-client==5.0.1

# ─── Document Loaders ───────────────────────────────────────────────────────────
pypdf==4.3.1
python-docx==1.1.2
unstructured==0.14.10
python-magic==0.4.27

# ─── Utilities ──────────────────────────────────────────────────────────────────
Pillow==10.4.0
python-decouple==3.8
structlog==24.4.0
sentry-sdk==2.10.0
gunicorn==22.0.0
uvicorn[standard]==0.30.3
httpx==0.27.0
pydantic==2.8.2
